+++ mcsniperpy/TODO.md
@@ -1,0 +1,17 @@
+# Todo
+
+- Account loading
+- Mojang account authentication
+- Snipe request
+- command line interface (the only way to run the sniper)
+
+> once all of that is done nonessential stuff can be worked on
+
+## Nonessential todo
+
+- Microsoft authentication
+- plugin system (possibly)
+
+## Other
+
+- Bug fixes when they come up
