+++ mcsniperpy/cleaned-code/readme.md
@@ -1,0 +1,3 @@
+# INFO
+
+This code was cleaned up by [AlvinoNguyen](https://github.com/AlvinoNguyen) and then updated to the new API by [zylarcher](https://github.com/zylarcher/).
