+++ mcsniperpy/mcsniperpy/util/naming_system.py
