+++ mcsniperpy/mcsniperpy/util/request_manager.py
