--- mcsniperpy/requirements.txt
@@ -1,4 +1,0 @@
-aiohttp
-colorama
-bs4
-requests
